requirements.txt
opencv-python
numpy
customtkinter
pillow
ultralytics